# doc-api.user.swagger.yaml
openapi: 3.0.3
info:
  title: Document Api NestJs
  description: |-
    This is a sample Pet Store Server based on the OpenAPI 3.0 specification.  You can find out more about
    Swagger at [https://swagger.io](https://swagger.io). In the third iteration of the pet store, we've switched to the design first approach!
    You can now help us improve the API whether it's by making changes to the definition itself or to the code.
    That way, with time, we can improve the API in general, and expose some of the new features in OAS3.

    _If you're looking for the Swagger 2.0/OAS 2.0 version of Petstore, then click [here](https://editor.swagger.io/?url=https://petstore.swagger.io/v2/swagger.yaml). Alternatively, you can load via the `Edit > Load Petstore OAS 2.0` menu option!_

    Some useful links:
    - [The Pet Store repository](https://github.com/swagger-api/swagger-petstore)
    - [The source API definition for the Pet Store](https://github.com/swagger-api/swagger-petstore/blob/master/src/main/resources/openapi.yaml)
  termsOfService: http://swagger.io/terms/
  contact:
    email: tungdev64@gmail.com
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  version: 1.0.11
externalDocs:
  description: Find out more about Swagger
  url: http://swagger.io
servers:
  - url: http://localhost:3000
tags:
  - name: auth
    description: API login register user
    externalDocs:
      description: login register user
      url: http://swagger.io
paths:
  /auth/login:
    post:
      tags:
        - auth
      summary: Đăng nhập
      description: Đăng nhập người dùng
      operationId: authentication
      requestBody:
        description: Thông tin người dùng
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Login'
        required: true
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessfulLogin'

        '400':
          description: Invalid user
        '403':
          description: User not found
        '422':
          description: Validation exception
      security:
        - petstore_auth:
            - write:user
            - read:user
  /user/me:
    get:
      tags:
        - user
      summary: Thông tin chi tiết
      description: Thông tin chi tiết người dùng
      operationId: user detail
      security:
        - BearerAuth: []
      responses:
        default:
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DetailUser'

components:
  schemas:
    Login:
      type: object
      properties:
        email:
          type: string
          example: tungdev64@gmail.com
        password:
          type: string
          example: 12345
    SuccessfulLogin:
      type: object
      properties:
        message:
          type: string
          example: 'Đăng nhập thành công'
        data:
          type: object
          properties:
            id:
              type: number
              example: 1
            email:
              type: string
              example: tungdev64@gmail.com
        token:
          type: object
          properties:
            accessToken:
              type: string
              example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJzdWIiOjEsImVtYWlsIjoidHVuZzFAZ21haWwuY29tIiwiaWF0IjoxNzEyNjQ3NTAyLCJleHAiOjE3MTI2NDgxMDJ9.KSdxhLpq7T_JEmaq5yMfDKl2QKjK7X7KM3fZlXRCZ04
    DetailUser:
      type: object
      properties:
        id:
          type: number
          example: 1
        email:
          type: string
          example: tungdev64@gmail.com

  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT